requirements.txt
torch
numpy
segment_anything
Pillow
opencv-python

dghs-imgutils[gpu]